# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Cigsmoke API Docs",
    "version": "1.0.0",
    "title": "Cigsmoke"
  },
  "basePath": "/",
  "paths": {
...
    "/checkouts": {
      "get": {
        "tags": [
          "checkout"
        ],
        "summary": "Get array of checkouts",
        "description": "",
        "operationId": "getCheckouts",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "addressId",
            "in": "query",
            "description": "Filtering checkouts by addressId",
            "required": false,
            "type": "string"
          },
          {
            "name": "basketId",
            "in": "query",
            "description": "Filtering checkouts by basketId",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sorting checkouts by property",
            "default": "basket",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Ordering results. ASC | DESC",
            "default": "DESC",
            "required": false,
            "type": "string"
          },
          {
            "type": "string",
            "in": "query",
            "name": "offset",
            "required": false,
            "default": 0,
            "description": "qty rows to skip for pagination"
          },
          {
            "type": "string",
            "in": "query",
            "name": "limit",
            "required": false,
            "default": 10,
            "description": "qty rows to take for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CheckoutResponse"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "checkout"
        ],
        "summary": "Create a checkout",
        "description": "",
        "operationId": "createCheckout",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CheckoutDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Checkout request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/checkouts/{checkoutId}": {
      "get": {
        "tags": [
          "checkout"
        ],
        "summary": "Find checkout by id",
        "operationId": "findCheckoutById",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "checkoutId",
            "in": "path",
            "description": "ID of checkout to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Checkout"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Checkout not found"
          }
        }
      },
      "put": {
        "tags": [
          "checkout"
        ],
        "summary": "Update the checkout",
        "description": "",
        "operationId": "updateCheckout",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "checkoutId",
            "in": "path",
            "description": "ID of checkout to return",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CheckoutDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful updated",
            "schema": {
              "$ref": "#/definitions/CheckoutWithoutJoins"
            }
          },
          "400": {
            "description": "Invalid Checkout request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Checkout not found"
          }
        }
      },
      "delete": {
        "tags": [
          "checkout"
        ],
        "summary": "Delete the checkout",
        "description": "",
        "operationId": "deleteCheckout",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "checkoutId",
            "in": "path",
            "description": "ID of checkout to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful deleted",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "comment": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Checkout not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "AddressWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "userId": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "zipCode": {
          "type": "string"
        }
      }
    },
...
    "BasketWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "userId": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "status": {
          "type": "string",
          "default": "New"
        }
      }
    },
...
    "Checkout": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "address": {
          "type": "object",
          "$ref": "#/definitions/AddressWithoutJoins"
        },
        "basket": {
          "type": "object",
          "$ref": "#/definitions/BasketWithoutJoins"
        },
        "comment": {
          "type": "string"
        }
      }
    },
    "CheckoutResponse": {
      "properties": {
        "rows": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Checkout"
          }
        },
        "length": {
          "type": "number"
        }
      }
    },
    "CheckoutWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "userId": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        }
      }
    },
    "CheckoutDTO": {
      "required": [
        "address",
        "basket"
      ],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "address": {
          "type": "string",
          "description": "Address ID"
        },
        "basket": {
          "type": "string",
          "description": "Basket ID"
        },
        "comment": {
          "type": "string"
        }
      }
    },
...
  }
}